+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Client.NewRequest: shallow-copy BaseURL to avoid retaining modifications
 
 ## [1.3.0] - 2017-02-28
 ### Added