+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Client.NewRequest: split query to avoid encoding "?" as "%3F" into path
 
 ## [1.3.1] - 2017-03-03
 ### Changed